Move AppOperations tests into a namespace
