+++ tests/Microsoft.DotNet.XHarness.Apple.Tests/AppOperations/AppInstallerTests.cs
@@ -15,7 +15,7 @@ using Microsoft.DotNet.XHarness.iOS.Shared.Hardware;
 using Moq;
 using Xunit;
 
-namespace Microsoft.DotNet.XHarness.Apple.Tests
+namespace Microsoft.DotNet.XHarness.Apple.Tests.AppOperations
 {
     public class AppInstallerTests : IDisposable
     {
