+++ tests/Microsoft.DotNet.XHarness.Apple.Tests/AppOperations/AppRunnerTests.cs
@@ -17,7 +17,7 @@ using Moq;
 using Xunit;
 
 #nullable enable
-namespace Microsoft.DotNet.XHarness.Apple.Tests
+namespace Microsoft.DotNet.XHarness.Apple.Tests.AppOperations
 {
     public class AppRunnerTests : AppRunTestBase
     {
