+++ tests/Microsoft.DotNet.XHarness.Apple.Tests/AppOperations/AppTesterTests.cs
@@ -19,7 +19,7 @@ using Moq;
 using Xunit;
 
 #nullable enable
-namespace Microsoft.DotNet.XHarness.Apple.Tests
+namespace Microsoft.DotNet.XHarness.Apple.Tests.AppOperations
 {
     public class AppTesterTests : AppRunTestBase
     {
